Handling of undefined classes in torch-free version of the package
